rudimentary and extremely hacky initial C code output (not usable at the moment, this is just a PoC (Proof of Concept)).
The trick is to allow the packer to write its binary data to file, then flush it to disk to make sure everything is there, and then employ the gup_io_reload() API to 'reload' that data back into a separate memory buffer.

Yes, hacky, but the original 'chainable calls' design for the compressor was never implemented in C++ as buf_write_nocrc() et al are not chainable: they go striaght to the std runtime library write calls, instead of referring back to the invoking archive class instance -- this COULD have been done, but have meant slightly more complex code at the time and consequently more writing and testing effort required. So the chaining idea remained an idea all these years...

Ameliorating argument: the DUMP mode outputs are not especially demanding re performance, so a flush+seek+load action is o-kay. sort of...