Adding a 2M+ large test file has born fruit: b0rk b0rk b0rk. Subtle failures when 'flipping' a file buffer from write to read mode (for depacking+crc_checking in dump mode!) and back again. We had to add a few new gup I/O APIs, nasty buggers:
    gup_result gup_io_switch_to_read_mode(buf_fhandle_t* file);
    gup_result gup_io_switch_to_write_mode(buf_fhandle_t* file);

Now these are used to 'properly' switch a write stream back to 'read mode' (suitable for use with the depacker code) and, once done, flip it back to write mode again. The *bug* that drove this was that the depacker COULD re-use the write buffer 'n all, but things got nutso as soon as that buffer was depleted and needed filling, i.e. b0rkb0rkb0rk for large(ish) files! The above APIs fix this by

1) flushing the buffers, if there's anything in there still.
2) *properly* set up write-mode and read-mode buffers while switching modes.
3) making sure we remain at the exact same file position (offset) while we're doing this.